modified track code

+++ bgspy/genome.py
@@ -389,7 +389,7 @@ def process_annotation(features, recmap, split_length=None):
     assert(map_pos.shape[0] == ranges.shape[0])
     print(f"done.")
     assert np.sum(np.isnan(map_pos)) == 0, "some interpolatd map positions are nan!"
-    rates = np.array(split_rates, dtype='float32')
+    rates = np.array(split_rates)
     assert np.sum(np.isnan(split_rates)) == 0, "some segment recombination rates are nan!"
     features = np.array([feature_map[x] for x in split_features])
     return Segments(ranges, rates, map_pos, features, feature_map, index)
